Refactor project structure, remove adapters, and consolidate CLI commands
