+++ src/rldk/__init__.py
@@ -2,16 +2,26 @@
 
 __version__ = "0.1.0"
 
+# Main public API - Core tracking functionality
+from .tracking import ExperimentTracker, TrackingConfig
+
+# Core functions from other modules
 from .ingest import ingest_runs
 from .diff import first_divergence
 from .determinism import check
-from .bisect import bisect_commits
 from .reward import health, RewardHealthReport
 from .evals import run, EvalResult
 
+# Additional functionality
+from .diff import bisect_commits  # Moved from bisect module
+
 __all__ = [
+    # Main public API
+    "ExperimentTracker",
+    "TrackingConfig",
+    # Core functions
     "ingest_runs",
-    "first_divergence",
+    "first_divergence", 
     "check",
     "bisect_commits",
     "health",
