--- src/rldk/artifacts/__init__.py
@@ -1,1 +1,0 @@
-"""Artifacts module for RL Debug Kit."""
