--- src/rldk/bisect/__init__.py
@@ -1,5 +1,0 @@
-"""Git bisect wrapper for finding regressions."""
-
-from .bisect import bisect_commits, BisectResult
-
-__all__ = ["bisect_commits", "BisectResult"]
