+++ src/rldk/forensics/__init__.py
@@ -6,6 +6,11 @@ from .gradient_norms_analyzer import GradientNormsAnalyzer, GradientNormsMetrics
 from .advantage_statistics_tracker import AdvantageStatisticsTracker, AdvantageStatisticsMetrics
 from .comprehensive_ppo_forensics import ComprehensivePPOForensics, ComprehensivePPOMetrics
 
+# Moved from artifacts module
+from .env_audit import audit_environment
+from .log_scan import scan_logs
+from .ckpt_diff import diff_checkpoints
+
 __all__ = [
     "scan_ppo_events",
     "KLScheduleTracker",
@@ -16,4 +21,7 @@ __all__ = [
     "AdvantageStatisticsMetrics",
     "ComprehensivePPOForensics",
     "ComprehensivePPOMetrics",
+    "audit_environment",
+    "scan_logs",
+    "diff_checkpoints",
 ]
