+++ src/rldk/ingest/__init__.py
@@ -1,5 +1,19 @@
-"""Ingest training runs from various sources."""
+"""Data ingestion from various sources."""
 
-from .ingest import ingest_runs, ingest_runs_to_events
+from .ingest import ingest_runs
 
-__all__ = ["ingest_runs", "ingest_runs_to_events"]
+# Adapters for different training log formats
+from .base import BaseAdapter
+from .trl import TRLAdapter
+from .openrlhf import OpenRLHFAdapter
+from .wandb import WandBAdapter
+from .custom_jsonl import CustomJSONLAdapter
+
+__all__ = [
+    "ingest_runs",
+    "BaseAdapter",
+    "TRLAdapter",
+    "OpenRLHFAdapter",
+    "WandBAdapter",
+    "CustomJSONLAdapter",
+]
