--- src/rldk/replay/__init__.py
@@ -1,10 +1,0 @@
-"""Seeded replay utility for training runs."""
-
-from rldk.replay.replay import (
-    replay,
-    ReplayReport,
-    _compare_metrics,
-    _prepare_replay_command,
-)
-
-__all__ = ["replay", "ReplayReport", "_compare_metrics", "_prepare_replay_command"]
