+++ src/rldk/reward/__init__.py
@@ -1,5 +1,6 @@
-"""Reward module for RL Debug Kit."""
+"""Reward model analysis and drift detection."""
 
 from .health_analysis import health, RewardHealthReport
+from .reward import generate_reward_card  # Moved from cards module
 
-__all__ = ["health", "RewardHealthReport"]
+__all__ = ["health", "RewardHealthReport", "generate_reward_card"]
